requirements.txt
Django==1.10.6
boto3==1.4.4
djangorestframework==3.5.4
